Add AI workflow guidance and maintenance follow-ups
