+++ .github/ISSUE_TEMPLATE/ai-task.md
@@ -1,0 +1,28 @@
+---
+name: AI task
+about: A scoped task for Codex, Claude, Antigravity, or another AI coding agent
+title: ""
+labels: codex
+assignees: ""
+---
+
+## Goal
+
+
+## Scope
+
+-
+
+## Constraints
+
+- Do not change secrets or external service settings.
+- Do not mutate production Sanity content unless explicitly requested.
+- Avoid unrelated refactors.
+
+## Validation
+
+-
+
+## Out Of Scope
+
+-
