+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,13 @@
+## Summary
+
+-
+
+## Validation
+
+- [ ] Ran the relevant build/check commands, or explained why not.
+- [ ] Updated generated Sanity schema/types when schema or query shape changed.
+- [ ] Checked production-sensitive effects such as SEO, redirects, secrets, and Sanity content.
+
+## Notes
+
+-
