+++ apps/studio/schemaTypes/documents/faq.ts
@@ -1,35 +1,13 @@
-import { defineType, type SlugIsUniqueFn } from 'sanity'
-
-const isUniqueSlugByLang: SlugIsUniqueFn = async (slug, context) => {
-  const { document, getClient } = context
-  const currentId = document?._id?.replace(/^drafts\./, '')
-  const draftId = currentId ? `drafts.${currentId}` : undefined
-  const language = (document as { language?: string } | undefined)?.language
-
-  if (!slug?.current || !language) return true
-
-  const client = getClient({ apiVersion: '2023-10-16' })
-  const query =
-    '*[_type == $type && slug.current == $slug && language == $language && !(_id in [$id, $draftId])][0]._id'
-  const params = {
-    type: document?._type,
-    slug: slug.current,
-    language,
-    id: currentId,
-    draftId,
-  }
-
-  const existingId = await client.fetch<string | null>(query, params)
-  return !existingId
-}
+import {defineType} from 'sanity'
+import {isUniqueSlugByLang} from '../utils/isUniqueSlugByLang'
 
 export default defineType({
   name: 'faq',
   title: 'FAQ',
   type: 'document',
   preview: {
-    select: { title: 'title', slug: 'slug.current', language: 'language' },
-    prepare({ title, slug, language }) {
+    select: {title: 'title', slug: 'slug.current', language: 'language'},
+    prepare({title, slug, language}) {
       const parts = [language, slug].filter(Boolean)
       return {
         title: title ?? 'Untitled',
@@ -44,11 +22,11 @@ export default defineType({
       readOnly: true,
       hidden: true,
     },
-    { name: 'title', type: 'string', title: 'Page Title' },
+    {name: 'title', type: 'string', title: 'Page Title'},
     {
       name: 'slug',
       type: 'slug',
-      options: { source: 'title', isUnique: isUniqueSlugByLang },
+      options: {source: 'title', isUnique: isUniqueSlugByLang},
       validation: (Rule) => Rule.required(),
     },
     {
@@ -65,11 +43,11 @@ export default defineType({
           type: 'object',
           title: 'FAQ Item',
           fields: [
-            { name: 'question', type: 'string', title: 'Question' },
-            { name: 'answer', type: 'blockContent', title: 'Answer' },
+            {name: 'question', type: 'string', title: 'Question'},
+            {name: 'answer', type: 'blockContent', title: 'Answer'},
           ],
           preview: {
-            select: { title: 'question' },
+            select: {title: 'question'},
           },
         },
       ],
